backend/requirements.txt
django==5.0.*
djangorestframework==3.15.*
django-cors-headers==4.3.*
django-environ==0.11.*
django-filter==23.5
django-cloudinary-storage==0.3.0
cloudinary==1.38.*
django-ckeditor-5==0.2.1
djangorestframework-simplejwt==5.3.*
pandas==2.2.*
numpy==1.26.*
matplotlib==3.8.*
scikit-learn==1.4.*
stripe==8.5.*
psycopg2-binary==2.9.*
gunicorn==21.2.*
whitenoise==6.6.*
pillow==10.2.*
python-dotenv==1.0.*
django-storages
boto3
fpdf2
